requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
scikit-learn
numpy
Pillow
scipy>=0.17.0
openpyxl
opencv-python
flask
imutils
xlrd